### 2017-02-27 #187 @jsta
I would be interested in reviewing. This technique looks fascinating and very useful!
