### 2017-02-27 #189 @whedon
[REVIEW]: spatial-efd: A spatial-aware implementation of elliptical Fourier analysis
**Submitting author:** @sgrieve (<a href="http://orcid.org/0000-0003-1893-7363">Stuart Grieve</a>)
**Repository:** <a href="https://github.com/sgrieve/spatial_efd" target ="_blank">https://github.com/sgrieve/spatial_efd</a>
**Version:** v1.0.3
