### 2017-02-28 sgrieve/spatial_efd#1 @jsta
joss review - minor typos etc
Hi - I am going to list any minor typos and inconsistencies that I find in my review in this issue thread.

https://github.com/openjournals/joss-reviews/issues/189
