### 2017-02-28 sgrieve/spatial_efd#1 @jsta
The usage section of the README says to use a file called 'Shapefile.shp` but this file does not exist. I tried 'example_data.shp' but it only has 5 polygons (not 20).
