### 2017-02-28 sgrieve/spatial_efd#1 @sgrieve
Thanks for spotting this! I have corrected the filenames and also provided the relative path to the shapefile to try and make things clearer. I have also corrected the indexing of example_data.shp, so that the example code now generates the figures included in the readme.
