### 2017-02-28 sgrieve/spatial_efd#1 @jsta
The README makes it sound like `spatial_efd.plotComparison(ax, coeffs, harmonic, x, y, rotation=0.)` should show a plot. I found that it does not produce a figure unless `InitPlot` is called first and `SavePlot` is called afterwards. People who are uneasy digging into the source code might not discover this?
