### 2017-02-28 sgrieve/spatial_efd#1 @jsta
I am having difficulty producing the results from the normalized example:

```
import spatial_efd
shp = spatial_efd.LoadGeometries('spatial_efd/spatial_efd/tests/example_data.shp')

harmonic = 20

x, y, centroid = spatial_efd.ProcessGeometryNorm(shp[1])
coeffs = spatial_efd.CalculateEFD(x, y, harmonic)
coeffs, rotation = spatial_efd.normalize_efd(coeffs, size_invariant=True)

ax = spatial_efd.InitPlot()
spatial_efd.plotComparison(ax, coeffs, harmonic, x, y, rotation=rotation)
spatial_efd.SavePlot(ax, harmonic, 'normtest', 'png')
```

![Imgur](http://i.imgur.com/6HsYRYY.png)
