### 2017-02-28 sgrieve/spatial_efd#1 @sgrieve
There was a mistake in the code in the readme, in order to get the shapefile and the Fourier ellipse to align, the normalize_efd method has to be called with size_invariant=False, which produces the correct figure.
![normtest_20](https://cloud.githubusercontent.com/assets/10617231/23415488/3a8f1252-fdd7-11e6-84f1-deea61a66fcf.png)

I have added an additional comment to highlight this within the code block with the aim of making this clearer to the user.
