### 2017-02-28 sgrieve/spatial_efd#1 @sgrieve
I have added the initplot and saveplot methods to the readme to better demonstrate the functionality of the plotcomparison method.
