### 2017-02-28 bdbaa38
Fixing more review issues
