+++ README.rst
@@ -147,8 +147,9 @@ Note that as this plotting is performed using ``matplotlib`` many other formatti
 To plot an overlay of a Fourier ellipse and the original shapefile data, a convenience function has been provided to streamline the coordinate processing required. To plot non-normalized coefficients:
 
 .. code-block:: python
-
+    ax = spatial_efd.InitPlot()
     spatial_efd.plotComparison(ax, coeffs, harmonic, x, y, rotation=0.)
+    spatial_efd.SavePlot(ax, harmonic, '/plots/myComparison', 'png')
 
 Which produces a figure like this:
 
@@ -164,7 +165,9 @@ And to plot normalized coefficients, where the data has been processed using the
 
 .. code-block:: python
 
-    coeffs, rotation = spatial_efd.normalize_efd(coeffs, size_invariant=True)
+    # size_invariant must be set to false if a normalized Fourier ellipse
+    # is to be plotted alongside the shapefile data
+    coeffs, rotation = spatial_efd.normalize_efd(coeffs, size_invariant=False)
     spatial_efd.plotComparison(ax, coeffs, harmonic, x, y, rotation=rotation)
 
 Which produces a figure like this:
