### 2017-02-28 sgrieve/spatial_efd#1 @jsta
OK, its working for me now.

It seems that one of the code blocks you edited is not rendered on the repo main page. Maybe because its missing a newline? I'm not too familiar with `rst` syntax.
