### 2017-02-28 sgrieve/spatial_efd#1 @jsta
I accidentally ran `saveShapefile` with `prj` set to a file that doesn't exist and there was no error thrown. This might be a worthwhile defensive programming test.
