### 2017-02-28 sgrieve/spatial_efd#1 @jsta
My last comment is more of style suggestion. I think it would be helpful to break-up the documentation into separate sections dealing with either normalized analyses or spatial analyses rather than switching back and forth.
