### 2017-02-28 #189 @jsta
@cMadan @sgrieve Looks good to me. I made several style suggestions in the repo issue thread (https://github.com/sgrieve/spatial_efd/issues/1) but overall I think this paper is good to go. :+1:
