### 2017-03-01 4244034
Changing saveshape code to warn users when prj files are missign
