### 2017-03-01 sgrieve/spatial_efd#1 @sgrieve
I have fixed the code block in the readme, you are correct, it needed a blank line.

`saveShapefile` already had error handling so that it would not crash if an incorrect filename was supplied. I have made this more clear to the user by adding warnings if an invalid file is supplied. As the prj file is optional I don't want to throw an error but hopefully the warning will alert the user that something has gone wrong.
