### 2017-03-01 ad69faf
fixing some typos in the readme
