### 2017-03-01 96bab83
Clarifying testing in readme
