+++ README.rst
@@ -76,7 +76,10 @@ Many of these tests make use of the ``example_data.shp`` file which is a shapefi
 Usage
 ----------
 
-Load a shapefile:
+Normalized Data
+~~~~~~~~~~~~~~~~~~~~~~
+
+The first step in using ``spatial_efd`` is always to load a shapefile:
 
 .. code-block:: python
 
@@ -87,17 +90,9 @@ This creates a shapefile object ``shp`` which contains the polygon geometries we
 
 .. code-block:: python
 
-    x, y, centroid = spatial_efd.ProcessGeometry(shp[1])
-
-This loads the geometry from the 2nd polygon within the shapefile into a list of x and a list of y coordinates. This method also computes the centroid of the polygon, which can be useful for later analysis.
-
-Note that the contents of x, y and centroid have not been normalized and so this method should be used where the original coordinates need to be preserved, for example if output to a shapefile is desired.
-
-In cases where the original coordinates are not needed, a different processing method can be called, to return the same data, but normalized, which can make comparisons between data from different locations simpler:
-
-.. code-block:: python
-
     x, y, centroid = spatial_efd.ProcessGeometryNorm(shp[1])
+
+This loads the geometry from the 2nd polygon within the shapefile into a list of x and a list of y coordinates. This method also computes the centroid of the polygon, which can be useful for later analysis. To make comparisons between data from different locations simpler, these data are normalized.
 
 If you already know how many harmonics you wish to compute this can be specified during the calculation of the Fourier coefficients:
 
@@ -121,20 +116,13 @@ Once the coefficients have been calculated they can be normalized following the 
 
     coeffs, rotation = spatial_efd.normalize_efd(coeffs, size_invariant=True)
 
-``size_invariant`` should be set to True (the default value) in most cases to normalize the coefficient values, allowing comparison between polygons of differing sizes. Set ``size_invariant`` to False if it is required to plot the Fourier ellipses alongside the input shapefiles, or if the Fourier ellipses are to be written to a shapefile.
+``size_invariant`` should be set to True (the default value) in most cases to normalize the coefficient values, allowing comparison between polygons of differing sizes. Set ``size_invariant`` to False if it is required to plot the Fourier ellipses alongside the input shapefiles, or if the Fourier ellipses are to be written to a shapefile. These techniques which apply to normalized data are outlined later in this document.
 
 A set of coefficients can be converted back into a series of x and y coordinates by performing an inverse transform, where the harmonic value passed in will be the harmonic reconstructed:
 
 .. code-block:: python
 
     xt, yt = spatial_efd.inverse_transform(coeffs, harmonic=harmonic)
-
-Again, if plotting the data alongside the original shapefile data, the locus of the coefficients must also be computed and passed as an argument to the inverse transform method:
-
-.. code-block:: python
-
-    locus = spatial_efd.calculate_dc_coefficients(x, y)
-    xt, yt = spatial_efd.inverse_transform(coeffs, harmonic=harmonic, locus=locus)
 
 Wrappers around some of the basic ``matplotlib`` functionality is provided to speed up the visualization of results:
 
@@ -148,25 +136,8 @@ This example generates an axis object, plots our transformed coordinates onto it
 
 Note that as this plotting is performed using ``matplotlib`` many other formatting options can be applied to the created axis object, to easily create publication ready plots.
 
-To plot an overlay of a Fourier ellipse and the original shapefile data, a convenience function has been provided to streamline the coordinate processing required. To plot non-normalized coefficients:
-
-.. code-block:: python
-
-    ax = spatial_efd.InitPlot()
[…]
