### 2017-03-01 sgrieve/spatial_efd#1 @sgrieve
I have restructured the readme to now have two explicit sections, one dealing with spatial coordinates and one with normalized coordinates, so hopefully the readme is more useful.  I have also fixed a few other typos I spotted. Since there are no other comments I am going to close this issue.

Thanks for the feedback!
