### 2017-03-02 #189 @cMadan
@jsta - awesome, thank you!
@sgrieve - let me know if you're set and I'll do a final pass to confirm that everything follows JOSS' guidelines
