+++ setup.py
@@ -8,9 +8,9 @@ def readme():
 
 
 setup(name='spatial_efd',
-      version='1.0.3',
+      version='1.0.4',
       description='Spatial elliptical fourier analysis',
-      url='http://github.com/sgrieve',
+      url='http://github.com/sgrieve/spatial-efd',
       long_description=readme(),
       keywords='GIS elliptical fourier analysis shapefile',
       classifiers=['Development Status :: 5 - Production/Stable',
@@ -19,7 +19,7 @@ setup(name='spatial_efd',
                    'Intended Audience :: Science/Research',
                    'Topic :: Scientific/Engineering :: GIS'],
       author='Stuart WD Grieve',
-      author_email='s.grieve@ed.ac.uk',
+      author_email='s.grieve@qmul.ac.uk',
       license='MIT',
       packages=['spatial_efd'],
       install_requires=['matplotlib>=2.0.0', 'numpy>=1.12.0',
