+++ docs/source/conf.py
@@ -62,7 +62,7 @@ author = u'Stuart W.D. Grieve'
 # The short X.Y version.
 version = u'1.0'
 # The full version, including alpha/beta/rc tags.
-release = u'1.0.3'
+release = u'1.0.4'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
