### 2017-03-02 #189 @sgrieve
@cMadan - Everything is done on my end. I have just incremented the version number and updated the package on pypi, so everything is good to go!
