### 2017-03-07 #189 @arfon
Many thanks for the review here @jsta and for editing this submission @cMadan ✨

@sgrieve - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00189 ⚡️ 🚀 💥
